Revert schema delete in test

+++ tests/e2e/Adapter/Scopes/GeneralTests.php
@@ -424,9 +424,7 @@ trait GeneralTests
 
     public function testSharedTablesTenantPerDocument(): void
     {
-        /** @var Database $database */
         $database = static::getDatabase();
-
         $sharedTables = $database->getSharedTables();
         $tenantPerDocument = $database->getTenantPerDocument();
         $namespace = $database->getNamespace();
@@ -437,13 +435,8 @@ trait GeneralTests
             return;
         }
 
-        try {
-            /**
-             * MirrorTest returning false $database->exists(__FUNCTION__)
-             */
+        if ($database->exists(__FUNCTION__)) {
             $database->delete(__FUNCTION__);
-        } catch (\Throwable $e) {
-
         }
 
         $database
